requirements.txt
streamlit
langchain_community
unstructured
python-dotenv
anthropic
chromadb
sentence-transformers
openai